cups-backend/requirements.txt
aiosqlite==0.17.0
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.112.2
h11==0.14.0
httptools==0.6.1
idna==3.8
iso8601==1.1.0
pillow==10.4.0
pycups==2.0.4
pydantic==2.8.2
pydantic_core==2.20.1
PyPDF2==3.0.1
pypika-tortoise==0.1.6
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
reportlab==4.2.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
tortoise-orm==0.21.6
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.23.0
websockets==13.0
